pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "fraetor"
version = "0.1.0"
description = "Voice input application"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-transcription>=1.0.0b4",
    "azure-cognitiveservices-speech>=1.48.2",
    "fastapi>=0.135.1",
    "google-genai>=1.71.0",
    "numpy>=2.4.3",
    "pydantic>=2.12.5",
    "sounddevice>=0.5.5",
    "sse-starlette>=3.3.3",
    "uvicorn[standard]>=0.42.0",
]

[tool.ruff]
target-version = "py313"
src = ["src", "tests"]
line-length = 88

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
    "ARG",   # flake8-unused-arguments
    "PT",    # flake8-pytest-style
    "PTH",   # flake8-use-pathlib
    "TCH",   # flake8-type-checking
    "RUF",   # ruff-specific rules
    "S",     # flake8-bandit (セキュリティ)
    "PL",    # pylint 互換 (convention/error/refactor/warning)
]
ignore = [
    "COM812",  # ruff formatter と競合
    "ISC001",  # ruff formatter と競合
    "PLR0913", # too-many-arguments (DI で引数が多くなるため緩和)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG", "N802", "S", "PLC2401", "PLR2004"]

[tool.mypy]
python_version = "3.13"
mypy_path = "."
packages = ["src", "tests"]
plugins = ["pydantic.mypy"]
strict = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "sounddevice"
ignore_missing_imports = true
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "azure.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google.*"
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
timeout = 10

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-timeout>=2.4.0",
    "ruff>=0.15.4",
]

[project.scripts]
fraetor = "src.__main__:main"

[tool.uv]
exclude-newer = "14 days"
exclude-newer-package = { google-genai = "3 days" }
